requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
langdetect
transformers
torch
numpy
pydantic
huggingface_hub